call focusOnIncomingCalls on first queue entry update
Without it, if your phone was already ringing from a call from the
switchboard queue before you logged with your xivo client, the current
call widget would not be in the correct state.

An another way to do it would be to check the number of queue entry
during the switchboard xlet initialization, but the model doesn't
contain this information yet.