fix segfault when server stop when popup opened
We always should use QPointer when we use exec() to open a new window.
Because the parent of the new window can be closed at any moment by the
cti server, this let the window in a awkward state.
Moreover, the static function of popup (QMessageBox, QFileDialog)
shouldn't be used for the same reason.