config: enable starttls for old configurations
if the last time the xivo client was used was in version 16.01 or lower,
starttls will be enabled. this change will only happen one in production since
the version will be greater or equal to 16.02 when the configuration file is
saved. for dev versions starttls will be re-enabled everytime we lauch the xivo
client until the 16.02 version is tagged.

I also added braces to make the indentation between the beginGroup/endGroup more
editor friendly.