do not store password as cleartext anymore
This is not more secure, but at least it's not cleartext.

We can't make this more secure: we absolutely need to be able to send
the real password to the server, because xivo-auth expects the real
password, e.g. when logging in via LDAP credentials.

Even using the most sophisticated encryption algorithms, we would
still need to store the encryption key and the attacker, provided he
can read the settings file, could always decrypt the password.