people: remove the check for enablexfer before doing a transfer
the enablexfer field of the user is used to add the tT options to the Dial
application. It was checked to avoid showing the transfer menu if the option was
not enabled since it would not have been possible to do the transfer. Now that
transfers are done using xivo-ctid-ng, the tT options are not required anymore
to do transfers.

Adding tT when not required for the call flow also has the effect of not
allowing calls to be in direct media with DTMF in rfc2833